WFLY-16150 Deprecate ServiceExecutor and relocate to wildfly-clustering-ee-spi module
