+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/clustering/ee/cache/main/module.xml
@@ -41,7 +41,6 @@
         <module name="org.wildfly.clustering.marshalling.api"/>
         <module name="org.wildfly.clustering.marshalling.protostream"/>
         <module name="org.wildfly.clustering.ee.spi"/>
-        <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.common"/>
         <module name="org.wildfly.security.elytron-private"/>
     </dependencies>
